requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
beautifulsoup4
supabase
